server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Tyler Lozano <tyler.m.lozano@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
markdown-it-py = "^0.6.0"
Flask = "^1.1.2"
Flask-JWT-Extended = "^3.25.0"
Flask-SQLAlchemy = "^2.4.4"
Flask-Cors = "^3.0.9"
requests = "^2.25.1"
black = "^20.8b1"
Flask-RESTful = "^0.3.8"
flask-restplus = "^0.13.0"
uWSGI = "^2.0.19"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.4"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
